Controladores

+++ src/main/java/com/example/invOp_Global/service/ArticuloServiceImpl.java
@@ -65,18 +65,25 @@ public class ArticuloServiceImpl extends BaseServiceImpl<Articulo,Long> implemen
 
     @Override
     @Transactional
-    public boolean darBajaArticulo(Long idArticulo) throws Exception{
+    public boolean darBajaArticulo(Long idArticulo) throws Exception {
         Articulo articulo = articuloRepository.findById(idArticulo).orElseThrow(() -> new Exception("Artículo no encontrado"));
-        List<OrdenCompra> ordenesPendientes = ordenCompraRepository.findOrdenCompraByEstadoAndArticulo("PENDIENTE",idArticulo);
-        List<OrdenCompra> ordenesEnCurso = ordenCompraRepository.findOrdenCompraByEstadoAndArticulo("EN-CURSO",idArticulo);
-        if (!ordenesPendientes.isEmpty()){
+        List<OrdenCompra> ordenesPendientes = ordenCompraRepository.findOrdenCompraByEstadoAndArticulo("PENDIENTE", idArticulo);
+        List<OrdenCompra> ordenesEnCurso = ordenCompraRepository.findOrdenCompraByEstadoAndArticulo("EN-CURSO", idArticulo);
+        if (!ordenesPendientes.isEmpty()) {
             throw new Exception("El artículo no se puede dar de baja porque tiene órdenes de compra pendientes");
         }
-        // Primero elimina los registros de ordencompradetalle relacionados
+        if (!ordenesEnCurso.isEmpty()) {
+            throw new Exception("El artículo no se puede dar de baja porque tiene órdenes de compra en curso");
+        }
         List<DetalleOrdenCompra> detalles = detalleOCService.findDetalleOCByArticulo(idArticulo);
-
         for (DetalleOrdenCompra detalle : detalles) {
             detalleOCRepository.delete(detalle);
+        }
+        ;
+        List<ProveedorArticulo> proveedorArticulos = proveedorArticuloService.findProveedoresByArticulo(idArticulo);
+        for( ProveedorArticulo linea : proveedorArticulos){
+            Long idLinea = linea.getId();
+            proveedorArticuloService.delete(idLinea);
         }
 
         articuloRepository.delete(articulo);
@@ -111,9 +118,7 @@ public class ArticuloServiceImpl extends BaseServiceImpl<Articulo,Long> implemen
     public int calcularStockSeguridad(Long articuloId){
         Articulo articulo = findById(articuloId);
         double valorZ = 1.64 ;
-        System.out.println("Entro a query");
         Double tiempoDemora = proveedorArticuloService.findProveedorArticuloByArticuloAndProveedor(articulo.getId(),articulo.getProveedorPred().getId()).getTiempoDemora();
-        System.out.println("salgo");
         Double tiempoRevision = articulo.getTiempoRevision();
         if (tiempoRevision == null){
             tiempoRevision = 0.0;
@@ -231,7 +236,7 @@ public class ArticuloServiceImpl extends BaseServiceImpl<Articulo,Long> implemen
     public void modificarIntervaloFijo(Long articuloId) throws Exception{
         Articulo articulo = articuloRepository.findById(articuloId).orElseThrow(() -> new Exception("Artículo no encontrado"));
             articulo.setCantidadMaxima(null);
-            articulo.setTiempoRevision(null);
+            articulo.setCantidadAPedir(null);
             articulo.setModeloInventario(ModeloInventario.LOTE_FIJO);
             articuloRepository.save(articulo);
     }
@@ -239,26 +244,24 @@ public class ArticuloServiceImpl extends BaseServiceImpl<Articulo,Long> implemen
     @Override
     public void modificarLoteFijo(Long articuloId) throws Exception{
         Articulo articulo = articuloRepository.findById(articuloId).orElseThrow(() -> new Exception("Artículo no encontrado"));
-            articulo.setModeloInventario(ModeloInventario.INTERVALO_FIJO);
             articulo.setLoteOptimo(null);
             articulo.setPuntoPedido(null);
+            articulo.setModeloInventario(ModeloInventario.INTERVALO_FIJO);
             articuloRepository.save(articulo);
     }
 
     @Override
     public void modificarModeloInventario(Long articuloId) throws Exception{
         Articulo articulo = articuloRepository.findById(articuloId).orElseThrow(() -> new Exception("Artículo no encontrado"));
-        System.out.println("Hola");
             if(articulo.getModeloInventario().equals(ModeloInventario.LOTE_FIJO)){
-                System.out.println("Hola1");
                 modificarLoteFijo(articuloId);
-                System.out.println("Hola2");
                 calculosIntervaloFijo(articuloId);
             }
             if(articulo.getModeloInventario().equals(ModeloInventario.INTERVALO_FIJO)){
                 modificarIntervaloFijo(articuloId);
                 calculosLoteFijo(articuloId);
             }
+            articuloRepository.save(articulo);
     }
 
     @Override
@@ -276,8 +279,8 @@ public class ArticuloServiceImpl extends BaseServiceImpl<Articulo,Long> implemen
     @Override
     public void calcularTodo(Long articuloId) throws Exception {
         Articulo articulo = articuloRepository.findById(articuloId).orElseThrow(() -> new Exception("Artículo no encontrado"));
-        articulo.setStockSeguridad(calcularStockSeguridad(articuloId));
         articulo.setDemandaAnual(demandaAnual(articuloId));
+        articulo.setStockSeguridad(calcularStockSeguridad(articuloId));
         if(articulo.getModeloInventario() == ModeloInventario.LOTE_FIJO){
             articulo.setPuntoPedido(calculoPuntoPedido(articuloId));
             articulo.setLoteOptimo(calculoLoteOptimo(articuloId));
